react-app: Move feature demo to dummy screen
refs #33
